adding CI changes
