--- quaddtype/.flake8
@@ -1,6 +1,0 @@
-[flake8]
-max-line-length = 100
-ignore = D107,D104,W503
-per-file-ignores = __init__.py:F401,tests/*:D100
-docstring-convention = numpy
-docstring_style = numpy
