+++ quaddtype/pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
 requires = [
-    "meson>=0.63.0",
+    "meson>=1.3.2",
     "meson-python",
     "patchelf",
     "wheel",
@@ -13,7 +13,7 @@ name = "quaddtype"
 description = "Quad (128-bit) float dtype for numpy"
 version = "0.0.1"
 readme = 'README.md'
-author = "Peyton Murray"
+author = "Swayam Singh"
 requires-python = ">=3.9.0"
 dependencies = [
     "numpy"
