+++ quaddtype/quaddtype/src/casts.cpp
