+++ quaddtype/quaddtype/src/quaddtype_main.c
@@ -1,63 +1,46 @@
 #include <Python.h>
 
-#define PY_ARRAY_UNIQUE_SYMBOL quaddtype_ARRAY_API
-#define PY_UFUNC_UNIQUE_SYMBOL quaddtype_UFUNC_API
+#define PY_ARRAY_UNIQUE_SYMBOL QuadPrecType_ARRAY_API
+#define PY_UFUNC_UNIQUE_SYMBOL QuadPrecType_UFUNC_API
 #define NPY_NO_DEPRECATED_API NPY_2_0_API_VERSION
+#define NPY_TARGET_VERSION NPY_2_0_API_VERSION
+
 #include "numpy/arrayobject.h"
-#include "numpy/ufuncobject.h"
 #include "numpy/dtype_api.h"
 
 #include "dtype.h"
-#include "umath.h"
 
 static struct PyModuleDef moduledef = {
         PyModuleDef_HEAD_INIT,
-        .m_name = "quaddtype_main",
-        .m_doc = "Quad (128-bit) floating point experimental numpy dtype",
+        .m_name = "_quaddtype_main",
+        .m_doc = "Quad (128-bit) floating point Data Type for Numpy",
         .m_size = -1,
 };
 
-// Initialize the python module
 PyMODINIT_FUNC
 PyInit__quaddtype_main(void)
 {
     import_array();
-    import_umath();
-
     PyObject *m = PyModule_Create(&moduledef);
-    if (m == NULL) {
-        PyErr_SetString(PyExc_ImportError, "Unable to create the quaddtype_main module.");
+    if (!m) 
+    {
         return NULL;
     }
 
-    PyObject *mod = PyImport_ImportModule("quaddtype");
-    if (mod == NULL) {
-        PyErr_SetString(PyExc_ImportError, "Unable to import the quaddtype module.");
+    if (init_quadprecision_scalar() < 0)
         goto error;
-    }
-    QuadScalar_Type = (PyTypeObject *)PyObject_GetAttrString(mod, "QuadScalar");
-    Py_DECREF(mod);
-    if (QuadScalar_Type == NULL) {
-        PyErr_SetString(PyExc_AttributeError,
-                        "Unable to find QuadScalar attribute in the "
-                        "quaddtype_main module.");
+    
+    if(PyModule_AddObject(m, "QuadPrecision", (PyObject *)&QuadPrecision_Type) < 0)
         goto error;
-    }
-    if (init_quad_dtype() < 0) {
-        PyErr_SetString(PyExc_AttributeError, "QuadDType failed to initialize.");
-        goto error;
-    }
-    if (PyModule_AddObject(m, "QuadDType", (PyObject *)&QuadDType) < 0) {
-        PyErr_SetString(PyExc_TypeError, "Failed to add QuadDType to the quaddtype_main module.");
-        goto error;
-    }
 
-    if (init_multiply_ufunc() == -1) {
-        PyErr_SetString(PyExc_TypeError, "Failed to initialize the quadscalar multiply ufunc.");
+    if(init_quadprec_dtype() < 0)
         goto error;
-    }
+
+    if(PyModule_AddObject(m, "QuadPrecDType", (PyObject *)&QuadPrecDType) < 0)
+        goto error;
 
     return m;
+    
 
 error:
     Py_DECREF(m);
