+++ quaddtype/quaddtype/src/scalar.c
