+++ quaddtype/quaddtype/src/scalar.h
