--- quaddtype/quaddtype/src/umath.h
@@ -1,7 +1,0 @@
-#ifndef _NPY_UFUNC_H
-#define _NPY_UFUNC_H
-
-int
-init_multiply_ufunc(void);
-
-#endif /*_NPY_UFUNC_H */
