--- quaddtype/tests/conftest.py
@@ -1,3 +1,0 @@
-import os
-
-os.environ["NUMPY_EXPERIMENTAL_DTYPE_API"] = "1"
