--- quaddtype2/pyproject.toml
@@ -1,25 +1,0 @@
-[build-system]
-requires = [
-    "meson>=1.3.2",
-    "meson-python",
-    "patchelf",
-    "wheel",
-    "numpy"
-]
-build-backend = "mesonpy"
-
-[project]
-name = "quaddtype"
-description = "Quad (128-bit) float dtype for numpy"
-version = "0.0.1"
-readme = 'README.md'
-author = "Swayam Singh"
-requires-python = ">=3.9.0"
-dependencies = [
-    "numpy"
-]
-
-[project.optional-dependencies]
-test = [
-    "pytest",
-]
