--- quaddtype2/quaddtype/__init__.py
@@ -1,1 +1,0 @@
-from ._quaddtype_main import QuadPrecDType, QuadPrecision
